+++ node/src/Transaction.ts
@@ -67,6 +67,7 @@ import {
     createRenameNX,
     createSAdd,
     createSCard,
+    createSInter,
     createSIsMember,
     createSMembers,
     createSMove,
@@ -703,6 +704,19 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public scard(key: string): T {
         return this.addAndReturn(createSCard(key));
+    }
+
+    /** Gets the intersection of all the given sets.
+     * When in cluster mode, all `keys` must map to the same hash slot.
+     * See https://valkey.io/docs/latest/commands/sinter/ for more details.
+     *
+     * @param keys - The `keys` of the sets to get the intersection.
+     *
+     * Command Response - A set of members which are present in all given sets.
+     * If one or more sets do not exist, an empty set will be returned.
+     */
+    public sinter(keys: string[]): T {
+        return this.addAndReturn(createSInter(keys), true);
     }
 
     /** Returns if `member` is a member of the set stored at `key`.