+++ node/tests/RedisClusterClient.test.ts
@@ -287,6 +287,7 @@ describe("RedisClusterClient", () => {
                 client.brpop(["abc", "zxy", "lkn"], 0.1),
                 client.smove("abc", "zxy", "value"),
                 client.renamenx("abc", "zxy"),
+                client.sinter(["abc", "zxy", "lkn"]),
                 // TODO all rest multi-key commands except ones tested below
             ];