+++ node/tests/SharedTests.ts
@@ -1172,6 +1172,52 @@ export function runBaseTests<Context>(config: {
     );
 
     it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `sinter test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key1 = `{key}-1-${uuidv4()}`;
+                const key2 = `{key}-2-${uuidv4()}`;
+                const non_existing_key = `{key}`;
+                const member1_list = ["a", "b", "c", "d"];
+                const member2_list = ["c", "d", "e"];
+
+                // positive test case
+                expect(await client.sadd(key1, member1_list)).toEqual(4);
+                expect(await client.sadd(key2, member2_list)).toEqual(3);
+                expect(await client.sinter([key1, key2])).toEqual(
+                    new Set(["c", "d"]),
+                );
+
+                // invalid argument - key list must not be empty
+                try {
+                    expect(await client.sinter([])).toThrow();
+                } catch (e) {
+                    expect((e as Error).message).toMatch(
+                        "ResponseError: wrong number of arguments",
+                    );
+                }
+
+                // non-existing key returns empty set
+                expect(await client.sinter([key1, non_existing_key])).toEqual(
+                    new Set(),
+                );
+
+                // non-set key
+                expect(await client.set(key2, "value")).toEqual("OK");
+
+                try {
+                    expect(await client.sinter([key2])).toThrow();
+                } catch (e) {
+                    expect((e as Error).message).toMatch(
+                        "Operation against a key holding the wrong kind of value",
+                    );
+                }
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
         `sismember test_%p`,
         async (protocol) => {
             await runTest(async (client: BaseClient) => {