+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Python: Added RENAMENX command ([#1492](https://github.com/aws/glide-for-redis/pull/1492))
 * Python: Added PFCOUNT command ([#1493](https://github.com/aws/glide-for-redis/pull/1493))
 * Python: Added PFMERGE command ([#1497](https://github.com/aws/glide-for-redis/pull/1497))
+* Node: Added SINTER command ([#1500](https://github.com/aws/glide-for-redis/pull/1500))
 
 ## 0.4.0 (2024-05-26)